+++ src/util/BinaryContainerSetup.js
@@ -364,7 +364,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp0.response) : xmlhttp0.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }
+
+            var XHR_buffer = xmlhttp0.response;
 
             var geoNode = binGeo;
             var attribTypeStr = geoNode._vf.indexType;  //"Uint16"
@@ -448,7 +454,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp.response) : xmlhttp.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }
+
+            var XHR_buffer = xmlhttp.response;
 
             var geoNode = binGeo;
             var attribTypeStr = geoNode._vf.coordType;
@@ -566,7 +578,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp1.response) : xmlhttp1.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }    
+
+            var XHR_buffer = xmlhttp1.response;
 
             var geoNode = binGeo;
             var i = 0;
@@ -665,7 +683,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp2.response) : xmlhttp2.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }
+
+            var XHR_buffer = xmlhttp2.response;
 
             var attribTypeStr = binGeo._vf.normalType;
             shape._webgl.normalType = x3dom.Utils.getVertexAttribType(attribTypeStr, gl);
@@ -730,7 +754,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp3.response) : xmlhttp3.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }    
+
+            var XHR_buffer = xmlhttp3.response;
 
             var attribTypeStr = binGeo._vf.texCoordType;
             shape._webgl.texCoordType = x3dom.Utils.getVertexAttribType(attribTypeStr, gl);
@@ -816,7 +846,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp4.response) : xmlhttp4.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }    
+
+            var XHR_buffer = xmlhttp4.response;
 
             var attribTypeStr = binGeo._vf.colorType;
             shape._webgl.colorType = x3dom.Utils.getVertexAttribType(attribTypeStr, gl);
@@ -878,7 +914,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp5.response) : xmlhttp5.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }
+
+            var XHR_buffer = xmlhttp5.response;
 
             var attribTypeStr = binGeo._vf.tangentType;
             shape._webgl.tangentType = x3dom.Utils.getVertexAttribType(attribTypeStr, gl);
@@ -940,7 +982,13 @@ x3dom.BinaryContainerLoader.setupBinGeo = function(shape, sp, gl, viewarea, curr
             if (!shape._webgl)
                 return;
 
-            var XHR_buffer = binGeo._vf.compressed == true ? x3dom.Utils.gunzip(xmlhttp6.response) : xmlhttp6.response;
+            if ( binGeo._vf.compressed )
+            {
+                x3dom.debug.logError( "x3dom 1.8.2+ do not support compressed BinaryGeometries anymore" );
+                return;
+            }
+
+            var XHR_buffer = xmlhttp6.response;
 
             var attribTypeStr = binGeo._vf.binormalType;
             shape._webgl.binormalType = x3dom.Utils.getVertexAttribType(attribTypeStr, gl);